feat: move target tools into tools package
