+++ .mage/magefile.go
@@ -6,9 +6,7 @@ import (
 	// mage:import
 	_ "github.com/einride/mage-tools/make"
 	// mage:import
-	_ "github.com/einride/mage-tools/golangci_lint"
-	// mage:import
-	_ "github.com/einride/mage-tools/goreview"
+	_ "github.com/einride/mage-tools/tools/golangci_lint"
 	"github.com/einride/mage-tools/tools"
 )
